# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bainder API",
    "description": "bainder API OpenAPI schema",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "epub",
              "pdf",
              "image",
              "text",
              "other"
            ]
          },
          "mimeType": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "sha256": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "sensitive": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploading",
              "processing",
              "processed",
              "failed"
            ]
          },
          "errorReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "mimeType",
          "originalFilename",
          "sizeBytes",
          "sha256",
          "title",
          "sensitive",
          "status",
          "errorReason",
          "createdAt",
          "updatedAt"
        ]
      },
      "DocumentStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploading",
              "processing",
              "processed",
              "failed"
            ]
          },
          "errorReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "errorReason"
        ]
      },
      "EpubBook": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "publishedDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "identifiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coverImage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "chapterCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "wordCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "documentId",
          "authors",
          "language",
          "description",
          "publisher",
          "publishedDate",
          "identifiers",
          "subjects",
          "coverImage",
          "chapterCount",
          "wordCount"
        ]
      },
...
      "EpubDetail": {
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/EpubBook"
          },
          "toc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpubTocItem"
            }
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpubChapterSummary"
            }
          }
        },
        "required": [
          "book",
          "toc",
          "chapters"
        ]
      },
      "EpubChapter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "href": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "html": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "wordCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "linear": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "documentId",
          "order",
          "href",
          "title",
          "html",
          "text",
          "wordCount",
          "linear"
        ]
      },
...
      "PdfDetail": {
        "type": "object",
        "properties": {
          "pdf": {
            "$ref": "#/components/schemas/PdfDocument"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PdfPageSummary"
            }
          }
        },
        "required": [
          "pdf",
          "pages"
        ]
      },
      "PdfPage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "pageNumber": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "text": {
            "type": "string"
          },
          "wordCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "documentId",
          "pageNumber",
          "text",
          "wordCount"
        ]
      },
      "ImageDocument": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string"
          },
          "width": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "height": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "format": {
            "type": "string",
            "enum": [
              "jpeg",
              "png",
              "webp",
              "gif",
              "heic",
              "tiff",
              "bmp",
              "unknown"
            ]
          }
        },
        "required": [
          "documentId",
          "width",
          "height",
          "format"
        ]
      },
      "TextDocument": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string"
          },
          "charset": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "charset",
          "text"
        ]
      },
...
    }
  },
  "paths": {
...
    "/documents": {
      "post": {
        "operationId": "document.create",
        "summary": "Upload a document",
        "description": "Multipart upload. Required field `file` (binary). Optional `sensitive` (boolean string). Returns the created document with status='processing'; processing runs asynchronously via Workflow.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "sensitive": {
                    "type": "string",
                    "enum": [
                      "true",
                      "false"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document created (processing started)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "413": {
            "description": "Upload exceeds size limit"
          },
          "415": {
            "description": "Unsupported file format"
          }
        }
      },
      "get": {
        "operationId": "document.list",
        "summary": "List documents",
        "responses": {
          "200": {
            "description": "All documents owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "operationId": "document.get",
        "summary": "Get a document",
        "responses": {
          "200": {
            "description": "Document metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      },
      "delete": {
        "operationId": "document.delete",
        "summary": "Delete a document",
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/status": {
      "get": {
        "operationId": "document.getStatus",
        "summary": "Get processing status",
        "description": "Lightweight payload for polling while a document is being processed.",
        "responses": {
          "200": {
            "description": "Processing status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/raw": {
      "get": {
        "operationId": "document.getRaw",
        "summary": "Stream the original document blob",
        "responses": {
          "200": {
            "description": "Original bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/assets/{name}": {
      "get": {
        "operationId": "document.getAsset",
        "summary": "Fetch a derived document asset by name",
        "responses": {
          "200": {
            "description": "Asset bytes",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/epub": {
      "get": {
        "operationId": "document.getEpubDetail",
        "summary": "Get EPUB book detail (book metadata + TOC + chapter summaries)",
        "responses": {
          "200": {
            "description": "EPUB detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found or wrong kind"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/epub/chapters/{order}": {
      "get": {
        "operationId": "document.getEpubChapter",
        "summary": "Get a single EPUB chapter by linear order",
        "responses": {
          "200": {
            "description": "Chapter content (cleaned HTML and plain text)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubChapter"
                }
              }
            }
          },
          "400": {
            "description": "Invalid chapter order"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Document or chapter not found"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "order",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/pdf": {
      "get": {
        "operationId": "document.getPdfDetail",
        "summary": "Get PDF detail (metadata + page summaries)",
        "responses": {
          "200": {
            "description": "PDF detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found or wrong kind"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/pdf/pages/{page}": {
      "get": {
        "operationId": "document.getPdfPage",
        "summary": "Get a single PDF page by page number",
        "responses": {
          "200": {
            "description": "Page text and metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid page number"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Document or page not found"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "page",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/image": {
      "get": {
        "operationId": "document.getImage",
        "summary": "Get image metadata (width, height, format)",
        "responses": {
          "200": {
            "description": "Image metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDocument"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found or wrong kind"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/documents/{id}/text": {
      "get": {
        "operationId": "document.getText",
        "summary": "Get decoded text content",
        "responses": {
          "200": {
            "description": "Text content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextDocument"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Not found or wrong kind"
          },
          "409": {
            "description": "Document not yet processed"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
...
  }
}